Move webPush code into lib/webPush
