+++ api/resolvers/notifications.js
@@ -3,7 +3,7 @@ import { decodeCursor, LIMIT, nextNoteCursorEncoded } from '../../lib/cursor'
 import { getItem, filterClause, whereClause, muteClause } from './item'
 import { getInvoice } from './wallet'
 import { pushSubscriptionSchema, ssValidate } from '../../lib/validate'
-import { replyToSubscription } from '../webPush'
+import { replyToSubscription } from '../../lib/webPush'
 import { getSub } from './sub'
 
 export default {
